+++ src/pumpkin/Utilities/boolean_variable.h
@@ -13,7 +13,7 @@ class BooleanVariable
 {
 public:
 	BooleanVariable(); // create an undefined variable
-	explicit BooleanVariable(int index); //index > 0 must be
+	explicit BooleanVariable(long index); //index > 0 must be
 	BooleanVariable &operator=(BooleanVariable);
 
 	bool IsUndefined() const; //returns true or false if the variable is considered undefined. Internally a special code_ for undefined variables is kept to distinguish it, i.e. the never-used index zero variable
@@ -23,7 +23,7 @@ public:
 
 	size_t ToPositiveInteger() const; //used for the VectorObjectIndexed
 
-	int index_;
+	long index_;
 };
 
 } //end Pumpkin namespace
